+++ src/Entities/Common/CommonUIFactory.cpp
@@ -2,7 +2,7 @@
 
 #include "Managers/Game/World.h"
 
-uint8_t CommonUIFactory::CreateUIButton(World &world, Vector2 pos, Vector2 size, uint8_t baseColor, uint8_t selectedColor, int layer)
+uint8_t CommonUIFactory::CreateUIButton(World &world, Vector2 pos, Vector2 size, uint16_t baseColor, uint16_t selectedColor, int layer)
 {
     uint8_t id = world.CreateEntity();
 
@@ -40,7 +40,7 @@ uint8_t CommonUIFactory::CreateUIButton(World &world, Vector2 pos, Vector2 size,
     return id;
 }
 
-uint8_t CommonUIFactory::CreateUIText(World &world, Vector2 pos, const char* text, uint8_t textColor, int layer)
+uint8_t CommonUIFactory::CreateUIText(World &world, Vector2 pos, const char* text, uint16_t textColor, int layer)
 {
     uint8_t id = world.CreateEntity();
 
@@ -72,7 +72,7 @@ uint8_t CommonUIFactory::CreateUIText(World &world, Vector2 pos, const char* tex
     return id;
 }
 
-uint8_t CommonUIFactory::CreateTimerText(World &world, Vector2 pos, uint8_t color, int layer)
+uint8_t CommonUIFactory::CreateTimerText(World &world, Vector2 pos, uint16_t color, int layer)
 {
     uint8_t id = world.CreateEntity();
