+++ src/Entities/Components/UI/UIButtonComponent.h
@@ -9,7 +9,7 @@ struct UIButtonComponent
 {
     bool isSelected = false;
 
-    uint8_t defaultColor;
-    uint8_t selectedColor;
+    uint16_t defaultColor;
+    uint16_t selectedColor;
     std::function<void(GameManager& manager)> OnClicked; //might be empty and just depach events
 };
